Added task insertion into database and engineer moving

+++ backend/src/types/dto.rs
@@ -53,7 +53,7 @@ pub struct AssignEngineerRequest {
 #[derive(Serialize, ToSchema)]
 pub struct AssignEngineerResponse {
     pub engineer_uuid: String,
-    pub time: f32,
+    pub time: u64,
     pub time_limit_exceeded: bool,
     pub route: Route,
 }
